WYYScrapy/requirements.txt
attrs==22.1.0
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
dnspython==2.2.1
fake-useragent==1.0.1
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.10.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
packaging==21.3
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
PyDispatcher==2.0.6
pymongo==4.3.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
schedule==1.1.0
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.12
w3lib==2.0.1
zipp==3.10.0
zope.interface==5.5.2
